cacher: encapsulate watchCacheStorage and maintain Law of Demeter
