+++ staging/src/k8s.io/apiserver/pkg/storage/cacher/cacher.go
@@ -1384,7 +1384,7 @@ func newErrWatcher(err error) *errWatcher {
 
 func (c *Cacher) ShouldDelegateExactRV(resourceVersion string, recursive bool) (delegator.Result, error) {
 	// Not Recursive is not supported unitl exact RV is implemented for WaitUntilFreshAndGet.
-	if !recursive || c.watchCache.storage.snapshots == nil {
+	if !recursive || !c.watchCache.storage.SnapshottingEnabled() {
 		return delegator.Result{ShouldDelegate: true}, nil
 	}
 	listRV, err := c.versioner.ParseResourceVersion(resourceVersion)
@@ -1396,7 +1396,7 @@ func (c *Cacher) ShouldDelegateExactRV(resourceVersion string, recursive bool) (
 
 func (c *Cacher) ShouldDelegateContinue(continueToken string, recursive bool) (delegator.Result, error) {
 	// Not Recursive is not supported unitl exact RV is implemented for WaitUntilFreshAndGet.
-	if !recursive || c.watchCache.storage.snapshots == nil {
+	if !recursive || !c.watchCache.storage.SnapshottingEnabled() {
 		return delegator.Result{ShouldDelegate: true}, nil
 	}
 	_, continueRV, err := storage.DecodeContinue(continueToken, c.resourcePrefix)
@@ -1418,7 +1418,7 @@ func (c *Cacher) shouldDelegateExactRV(rv uint64) (delegator.Result, error) {
 			ShouldDelegate: !delegator.ConsistentReadSupported(),
 		}, nil
 	}
-	_, canServe := c.watchCache.storage.snapshots.GetLessOrEqual(rv)
+	canServe := c.watchCache.storage.CanServeExactRV(rv)
 	return delegator.Result{
 		ShouldDelegate: !canServe,
 	}, nil
