+++ staging/src/k8s.io/apiserver/pkg/storage/cacher/watch_cache_test.go
@@ -24,7 +24,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	v1 "k8s.io/api/core/v1"
@@ -1309,133 +1308,6 @@ func TestHistogramCacheReadWait(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestWatchCacheStorageSnapshots(t *testing.T) {
-	featuregatetesting.SetFeatureGateDuringTest(t, utilfeature.DefaultFeatureGate, features.ListFromCacheSnapshot, true)
-
-	keyFunc := func(obj runtime.Object) (string, error) {
-		return obj.(*mockObject).key, nil
-	}
-
-	indexers := &cache.Indexers{}
-	s := newWatchCacheStorage(keyFunc, indexers)
-
-	assert.True(t, s.snapshottingEnabled.Load(), "Expected snapshotting to be enabled when feature gate is active")
-
-	_, err := s.GetExactSnapshotLocked(100)
-	require.Error(t, err, "Expected empty cache to not include any snapshots")
-
-	t.Log("Test cache on rev 100")
-	elem1 := &store.Element{Key: "foo", Object: &mockObject{key: "foo", val: "100"}}
-	require.NoError(t, s.UpdateStoreLocked(watch.Added, elem1, 100))
-
-	elem2 := &store.Element{Key: "foo", Object: &mockObject{key: "foo", val: "200"}}
-	require.NoError(t, s.UpdateStoreLocked(watch.Modified, elem2, 200))
-
-	elem3 := &store.Element{Key: "foo", Object: &mockObject{key: "foo", val: "300"}}
-	require.NoError(t, s.UpdateStoreLocked(watch.Deleted, elem3, 300))
-
-	t.Log("Test cache on rev 100")
-	_, err = s.GetExactSnapshotLocked(99)
-	require.Error(t, err, "Expected store to not include rev 99")
-
-	snap100, err := s.GetExactSnapshotLocked(100)
-	require.NoError(t, err)
-	elements, err := snap100.OrderedListPrefix("", "")
-	require.NoError(t, err)
-	assert.Len(t, elements, 1)
-	assert.Equal(t, &mockObject{key: "foo", val: "100"}, elements[0].(*store.Element).Object)
-
-	t.Log("Compact snapshots to remove rev 100")
-	s.CompactSnapshotsLocked(200)
-	_, err = s.GetExactSnapshotLocked(100)
-	require.Error(t, err, "Expected compacted snapshot at 100 to be deleted")
-
-	t.Log("Test cache on rev 200")
-	snap200, err := s.GetExactSnapshotLocked(200)
-	require.NoError(t, err)
-	elements, err = snap200.OrderedListPrefix("", "")
-	require.NoError(t, err)
-	assert.Len(t, elements, 1)
-	assert.Equal(t, &mockObject{key: "foo", val: "200"}, elements[0].(*store.Element).Object)
-
-	t.Log("Test cache on rev 300")
-	snap300, err := s.GetExactSnapshotLocked(300)
-	require.NoError(t, err)
-	elements, err = snap300.OrderedListPrefix("", "")
-	require.NoError(t, err)
-	assert.Empty(t, elements)
-
-	t.Log("Test cache on rev 400")
-	elem4 := &store.Element{Key: "foo", Object: &mockObject{key: "foo", val: "400"}}
-	require.NoError(t, s.UpdateStoreLocked(watch.Added, elem4, 400))
-
-	snap400, err := s.GetExactSnapshotLocked(400)
-	require.NoError(t, err)
-	elements, err = snap400.OrderedListPrefix("", "")
-	require.NoError(t, err)
-	assert.Len(t, elements, 1)
-	assert.Equal(t, &mockObject{key: "foo", val: "400"}, elements[0].(*store.Element).Object)
-
-	t.Log("Compact snapshots to simulate cache capacity downsize")
-	s.CompactSnapshotsLocked(500)
-	_, err = s.GetExactSnapshotLocked(499)
-	require.Error(t, err, "Expected compacted snapshots below 500 to be deleted")
-
-	t.Log("Test cache on rev 500")
-	elem5 := &store.Element{Key: "foo", Object: &mockObject{key: "foo", val: "500"}}
-	require.NoError(t, s.UpdateStoreLocked(watch.Modified, elem5, 500))
-
-	snap500, err := s.GetExactSnapshotLocked(500)
-	require.NoError(t, err)
-	elements, err = snap500.OrderedListPrefix("", "")
-	require.NoError(t, err)
-	assert.Len(t, elements, 1)
-	assert.Equal(t, &mockObject{key: "foo", val: "500"}, elements[0].(*store.Element).Object)
-
-	t.Log("Test cache on rev 600")
-	elem6 := &store.Element{Key: "foo", Object: &mockObject{key: "foo", val: "600"}}
-	require.NoError(t, s.UpdateStoreLocked(watch.Modified, elem6, 600))
-
-	snap600, err := s.GetExactSnapshotLocked(600)
-	require.NoError(t, err)
-	elements, err = snap600.OrderedListPrefix("", "")
-	require.NoError(t, err)
-	assert.Len(t, elements, 1)
-	assert.Equal(t, &mockObject{key: "foo", val: "600"}, elements[0].(*store.Element).Object)
-
-	t.Log("Replace cache to remove history")
-	_, err = s.GetExactSnapshotLocked(500)
-	require.NoError(t, err, "Confirm that cache stores history before replace")
-
-	err = s.ReplaceLocked([]interface{}{
-		&store.Element{Key: "foo", Object: &mockObject{key: "foo", val: "600"}},
-	}, "700", 700)
-	require.NoError(t, err)
-
-	_, err = s.GetExactSnapshotLocked(500)
-	require.Error(t, err, "Expected replace to remove history")
-	_, err = s.GetExactSnapshotLocked(600)
-	require.Error(t, err, "Expected replace to remove history")
-
-	t.Log("Test cache on rev 700")
-	snap700, err := s.GetExactSnapshotLocked(700)
-	require.NoError(t, err)
-	elements, err = snap700.OrderedListPrefix("", "")
-	require.NoError(t, err)
-	assert.Len(t, elements, 1)
-	assert.Equal(t, &mockObject{key: "foo", val: "600"}, elements[0].(*store.Element).Object)
-}
-
-type mockObject struct {
-	runtime.Object
-	key string
-	val string
-}
-
-func (m *mockObject) DeepCopyObject() runtime.Object {
-	return &mockObject{key: m.key, val: m.val}
 }
 
 func TestWatchCacheSnapshotConcurrency(t *testing.T) {
